+++ syscall/samples/sched/policy.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <unistd.h>
 
-#include "helpers.h"
+#include "sched_helpers.h"
 
 int main(void)
 {